Issue #30, organized models from compound-data
