+++ src/app/_models/WDQSData.ts
@@ -1,0 +1,4 @@
+export interface WDQSData {
+  head: {vars: Array<string>};
+  results: {bindings: Array<Object>}
+}
