+++ src/app/_models/vendor-data/GVKData.ts
@@ -1,0 +1,12 @@
+export interface GVKData {
+  hvac_id: string;
+  gvk_id: string;
+  drug_name: string;
+  phase: Array<string>;
+  roa: Array<string>;
+  category: Array<string>;
+  mechanism: Array<string>;
+  smiles: string;
+  synonyms: Array<string>;
+  ikey: string;
+}
