+++ src/app/_models/vendor-data/assay-data.ts
@@ -1,0 +1,10 @@
+export interface AssayData {
+  id: string;
+  ac50: string;
+  datamode: string
+  assay_title: string;
+  smiles: string;
+  ikey: string;
+  wikidata: string;
+  pubchem_cid: string;
+}
