+++ src/app/_models/vendor-data/index.ts
@@ -1,0 +1,5 @@
+export * from './GVKData';
+export * from './integrity-data';
+export * from './informa-data';
+export * from './assay-data';
+export * from './vendor-data';
