+++ src/app/_models/vendor-data/informa-data.ts
@@ -1,0 +1,10 @@
+export interface InformaData {
+  drug_name: string;
+  phase: Array<string>;
+  highest_phase: string;
+  mechanism: Array<string>;
+  smiles: string;
+  ikey: string;
+  wikidata: string;
+  pubchem_cid: string;
+}
