+++ src/app/_models/vendor-data/vendor-data.ts
@@ -1,0 +1,9 @@
+import { GVKData, IntegrityData, InformaData, AssayData } from './index';
+
+export interface VendorData {
+  ikey: string;
+  gvk: GVKData;
+  integrity: IntegrityData;
+  informa: InformaData;
+  assay: Array<AssayData>;
+}
