+++ src/app/compound-data/compound-data.component.html
@@ -2,7 +2,7 @@
 
 
   <div class="row-fluid pull-right">
-    <a href="#search">
+    <a (click)="backClick()">
       << return to search</a>
   </div>
