+++ src/app/compound-data/compound-data.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit, forwardRef, Inject, Injectable, Input} from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {Http, Response} from "@angular/http";
+import { Location } from '@angular/common';
 
 import {WDQService} from "../compound-search/compound-search.component"
 // import { InteractionTableDataService } from "../interaction-table/interaction-table.component"
@@ -11,68 +12,9 @@ import {
   HttpClient, HttpErrorResponse, HttpEventType, HttpHeaders, HttpParams, HttpRequest,
   HttpResponse
 } from "@angular/common/http";
+
+import { AssayData, GVKData, IntegrityData, InformaData, VendorData, WDQSData } from '../_models/index';
 import {environment} from "../../environments/environment";
-
-
-export interface WDQSData {
-  head: {vars: Array<string>};
-  results: {bindings: Array<Object>}
-}
-
-export interface VendorData {
-  ikey: string;
-  gvk: GVKData;
-  integrity: IntegrityData;
-  informa: InformaData;
-  assay: Array<AssayData>;
-}
-
-export interface GVKData {
-  hvac_id: string;
-  gvk_id: string;
-  drug_name: string;
-  phase: Array<string>;
-  roa: Array<string>;
-  category: Array<string>;
-  mechanism: Array<string>;
-  smiles: string;
-  synonyms: Array<string>;
-  ikey: string;
-}
-
-export interface IntegrityData {
-  id: number;
-  smiles: string;
-  drug_name: string;
-  phase: Array<string>;
-  category: Array<string>;
-  mechanism: Array<string>;
-  ikey: string;
-  wikidata: string;
-  pubchem_cid: string;
-}
-
-export interface InformaData {
-  drug_name: string;
-  phase: Array<string>;
-  highest_phase: string;
-  mechanism: Array<string>;
-  smiles: string;
-  ikey: string;
-  wikidata: string;
-  pubchem_cid: string;
-}
-
-export interface AssayData {
-  id: string;
-  ac50: string;
-  datamode: string
-  assay_title: string;
-  smiles: string;
-  ikey: string;
-  wikidata: string;
-  pubchem_cid: string;
-}
 
 
 @Component({
@@ -149,7 +91,8 @@ export class CompoundDataComponent implements OnInit {
     private http: Http,
     private http2: HttpClient,
     // private interactionTableDataService: InteractionTableDataService,
-    private cidService: CIDService
+    private cidService: CIDService,
+    private _location: Location
   ) {
     route.params.subscribe(params => {
       this.qid = params['qid'];
@@ -224,6 +167,10 @@ export class CompoundDataComponent implements OnInit {
     //
     //   }
     // );
+  }
+
+  backClick() {
+    this._location.back();
   }
 
   buildData(): void {
